pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-commons"
version = "0.1.0"
description = "Shared data access + LLM client library for the MSARN / Companies_News / Idee_Scraping ecosystem."
readme = "README.md"
requires-python = ">=3.10"  # 3.12 on Mac/Railway; 3.10 tolerated in CI sandbox
authors = [{ name = "Andreas Doepfert" }]
license = { text = "Proprietary" }

# Keep this list MINIMAL. research_commons is a thin shared layer; heavy ML
# dependencies (torch, sklearn, optuna, streamlit, etc.) MUST stay in the
# individual project requirements.* files, not here.
dependencies = [
    "psycopg2-binary>=2.9,<3.0",
    "pandas>=2.2,<3.0",
    "anthropic>=0.39,<1.0",
    "google-generativeai>=0.8,<1.0",  # Gemini fallback
    "httpx>=0.27,<1.0",
    "python-dotenv>=1.0,<2.0",
    "pydantic>=2.7,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["research_commons*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
